IgaFormation

Index

IgaBase.QuadratureRule

IgaBase.QuadratureRule

IgaFormation.Element

IgaFormation.ElementAccessor

IgaFormation.Elements

IgaFormation.PatchAccessor

IgaFormation.UnivariatePatchData

IgaFormation.QuadraturePoints

IgaFormation.QuadraturePoints

IgaFormation.QuadratureWeights

IgaFormation.QuadratureWeights

IgaFormation.TestFunctions

IgaFormation.TestFunctions

IgaFormation.TestIndices

IgaFormation.TestIndices

IgaFormation.TrialFunctions

IgaFormation.TrialFunctions

IgaFormation.TrialIndices

IgaFormation.TrialIndices

IgaFormation.coface

IgaFormation.pre_allocate_arrays

IgaFormation.pre_allocate_arrays

IgaFormation.pre_allocate_arrays

IgaFormation.pre_allocate_arrays

IgaFormation.sumfact!

IgaFormation.sumfact!

IgaFormation.sumfact!

IgaFormation.sumfact!

IgaFormation.QuadraturePoints — Method

QuadraturePoints(acc::ElementAccessor{Dim}, element::Element{Dim}) where Dim

Get the quadrature points on this element.

IgaFormation.QuadraturePoints — Method

QuadraturePoints(acc::PatchAccessor{Dim}) where Dim

Get the quadrature points on this element.

IgaFormation.QuadratureWeights — Method

QuadratureWeights(acc::ElementAccessor{Dim}, element::Element{Dim}) where Dim

Get the quadrature weights on this element.

IgaFormation.QuadratureWeights — Method

QuadratureWeights(acc::PatchAccessor{Dim}) where Dim

Get the quadrature weights on this element.

IgaFormation.TestFunctions — Method

TestFunctions(acc::ElementAccessor{Dim}, element::Element{Dim}) where Dim

Get the test functions on an element as a KroneckerProduct matrix.

IgaFormation.TestFunctions — Method

TestFunctions(acc::PatchAccessor{Dim}) where Dim

Get the test functions on the patch as a KroneckerProduct matrix.

IgaFormation.TestIndices — Method

TestIndices(acc::PatchtAccessor)

Get global indices of the test functions.

IgaFormation.TestIndices — Method

TestIndices(acc::ElementAccessor{Dim}, element::Element{Dim})

Get global indices of the test functions on an element.

IgaFormation.TrialFunctions — Method

TrialFunctions(acc::ElementAccessor{Dim}, element::Element{Dim})

Get the trial functions on an element as a KroneckerProduct matrix.

IgaFormation.TrialFunctions — Method

TrialFunctions(acc::PatchAccessor)

Get the trial functions on the patch as a KroneckerProduct matrix.

IgaFormation.TrialIndices — Method

TrialIndices(acc::PatchAccessor)

Get global indices of the trial functions.

IgaFormation.TrialIndices — Method

TrialIndices(acc::ElementAccessor{Dim}, element::Element{Dim})

Get global indices of the trial functions on an element.

IgaFormation.coface — Method

coface(element::Element)

If element is on the boundary then element.parent is a view and this function returns its single co-face.

IgaFormation.pre_allocate_arrays — Method

pre_allocate_arrays(T, nquad::NTuple{3,Int}, m::NTuple{3,Int}, n::NTuple{3,Int})

Preallocate arrays of element type==T for use in sum factorization of matrices.

IgaFormation.pre_allocate_arrays — Method

pre_allocate_arrays(T, nquad::NTuple{3,Int}, m::NTuple{3,Int})

Preallocate arrays of element type==T for use in sum factorization of vectors.

IgaFormation.pre_allocate_arrays — Method

pre_allocate_arrays(T, nquad::NTuple{2,Int}, m::NTuple{2,Int}, n::NTuple{2,Int})

Preallocate arrays of element type==T for use in sum factorization of matrices.

IgaFormation.pre_allocate_arrays — Method

pre_allocate_arrays(T, nquad::NTuple{2,Int}, m::NTuple{2,Int})

Preallocate arrays of element type==T for use in sum factorization of vectors.

IgaFormation.sumfact! — Method

sumfact!(A::Tuple{<:Array,<:Array,<:Array}, C, testfuns)

Apply sum factorization. A and B are allocated using pre_allocate_arrays(...). testfuns and trialfuns are tuples of arrays that represent the test and trial functions evaluated at the quadrature nodes. It is assumed that the quadrature weights have been incorporated into the test functions.

IgaFormation.sumfact! — Method

sumfact!(A::Tuple{<:Array,<:Array,<:Array}, B, C, testfuns, trialfuns)

Apply sum factorization. A and B are allocated using pre_allocate_arrays(...). testfuns and trialfuns are tuples of arrays that represent the test and trial functions evaluated at the quadrature nodes. It is assumed that the quadrature weights have been incorporated into the test functions

IgaFormation.sumfact! — Method

sumfact!(A::Tuple{<:Array,<:Array}, C, testfuns)

Apply sum factorization. A is pre-allocated using pre_allocate_rhs_arrays(...). testfuns is a tuple of matrices that represent the test functions evaluated at the quadrature nodes. It is assumed that the quadrature weights have been incorporated into the test functions.

IgaFormation.sumfact! — Method

sumfact!(A::Tuple{<:Array,<:Array}, B, C, testfuns, trialfuns)

Apply sum factorization. A and B are allocated using pre_allocate_arrays(...). testfuns and trialfuns are tuples of arays that represent the test and trial functions evaluated at the quadrature nodes. It is assumed that the quadrature weights have been incorporated into the test functions.

IgaBase.QuadratureRule — Method

QuadratureRule(acc::ElementAccessor{Dim}, element::Element{Dim}) where Dim

Get the quadrature rule on this element.

IgaBase.QuadratureRule — Method

QuadratureRule(acc::PatchAccessor{Dim}) where Dim

Get the quadrature rule on the patch.

IgaFormation.Element — Type

Element{Dim, T}

Dim-dimensional element storing its domain of definition as a Cartesian{Dim, Interval{T}} and its element number as a CartesianIndex{Dim}

IgaFormation.ElementAccessor — Type

ElementAccessor{Dim, X<:IncreasingSequence, Data}

Datastructure that provides access to element data such as the element domain, element trialfunctions, element testfunctions and element quadrature rule.

IgaFormation.Elements — Type

Elements(partition::CartesianProduct{Dim, <:IncreasingSequence{<:Real}})

Return an iterator over all elements in a partition.

IgaFormation.PatchAccessor — Type

PatchAccessor{Dim, X, ..., Data}

Datastructure that provides access to element data such as the element domain, element trialfunctions, element testfunctions and element quadrature rule.

IgaFormation.UnivariatePatchData — Type

UnivariatePatchData{T<:Real, Q<:AbstractQuadrule{1}}

Datastructure that stores all the univariate patch data. This data is conveniently accessed by an element accessor.